more meta updates and property renaming
